Updated code and move test cases in tests.rs
