--- pallets/template/src/mock.rs
@@ -1,2 +1,0 @@
-// Setup mock runtime for testing this pallet
-#![cfg(test)]
